+++ example/lib/bodies/sequential.dart
@@ -1,12 +1,6 @@
 import 'package:event_queue/event_queue.dart';
 import 'package:flutter/material.dart';
 
-enum _EventType {
-  event1,
-  event2,
-  event3,
-}
-
 class SequentialBody extends StatefulWidget {
   const SequentialBody({super.key});
 
@@ -15,11 +9,7 @@ class SequentialBody extends StatefulWidget {
 }
 
 class _SequentialBodyState extends State<SequentialBody> {
-  final _queue = SequentialEventQueue<_EventType>(
-    allowAsyncFor: (e1, e2) => [e1, e2].every(
-      (e) => e == _EventType.event2 || e == _EventType.event3,
-    ),
-  );
+  final _queue = EventQueue.sequential();
 
   @override
   void dispose() {
@@ -34,16 +24,8 @@ class _SequentialBodyState extends State<SequentialBody> {
         mainAxisSize: MainAxisSize.min,
         children: [
           TextButton(
-            onPressed: _fire1,
-            child: const Text('Fire1'),
-          ),
-          TextButton(
-            onPressed: _fire2,
-            child: const Text('Fire2'),
-          ),
-          TextButton(
-            onPressed: _fire3,
-            child: const Text('Fire3'),
+            onPressed: _fire,
+            child: const Text('Fire'),
           ),
           TextButton(
             onPressed: () => _queue.clear(),
@@ -54,21 +36,9 @@ class _SequentialBodyState extends State<SequentialBody> {
     );
   }
 
-  void _fire1() => _queue(() async {
-        print('fire10');
+  void _fire() => _queue(() async {
+        print('fire0');
         await Future.delayed(const Duration(seconds: 1));
-        print('fire11');
-      }, eventId: _EventType.event1);
-
-  void _fire2() => _queue(() async {
-        print('fire20');
-        await Future.delayed(const Duration(seconds: 1));
-        print('fire21');
-      }, eventId: _EventType.event2);
-
-  void _fire3() => _queue(() async {
-        print('fire30');
-        await Future.delayed(const Duration(seconds: 1));
-        print('fire31');
-      }, eventId: _EventType.event3);
+        print('fire1');
+      });
 }