+++ example/lib/main.dart
@@ -1,3 +1,4 @@
+import 'package:event_queue_example/bodies/async.dart';
 import 'package:event_queue_example/bodies/custom_priority.dart';
 import 'package:event_queue_example/bodies/droppable.dart';
 import 'package:event_queue_example/bodies/sequential.dart';
@@ -10,11 +11,12 @@ void main() {
 }
 
 enum CaseStudy {
+  sequential,
   droppable,
   singleElement,
-  sequential,
-  syncAsync,
   customPriority,
+  async,
+  syncAsync,
 }
 
 class _Page extends StatelessWidget {
@@ -31,42 +33,19 @@ class _Page extends StatelessWidget {
             tabs: CaseStudy.values.map((e) => Tab(text: e.name)).toList(),
           ),
         ),
-        body: const TabBarView(
-          children: [
-            DroppableBody(),
-            SingleElementBody(),
-            SequentialBody(),
-            SyncAsyncBody(),
-            CustomPriorityBody(),
-          ],
+        body: TabBarView(
+          children: CaseStudy.values
+              .map((e) => switch (e) {
+                    CaseStudy.sequential => const SequentialBody(),
+                    CaseStudy.droppable => const DroppableBody(),
+                    CaseStudy.singleElement => const SingleElementBody(),
+                    CaseStudy.customPriority => const CustomPriorityBody(),
+                    CaseStudy.async => const AsyncBody(),
+                    CaseStudy.syncAsync => const SyncAsyncBody(),
+                  })
+              .toList(),
         ),
       ),
     );
   }
 }
-
-// void _fire1() => _queue(() async {
-//         print('fire10');
-//         await Future.delayed(const Duration(seconds: 1));
-//         // if (!_queue.isEmpty) return;
-//         print('fire11');
-//       }, eventId: EventType.event1);
-
-//   void _fire2() => _queue(() async {
-//         print('fire20');
-//         await Future.delayed(const Duration(seconds: 1));
-//         print('fire21');
-//       }, eventId: EventType.event2);
-
-//   void _fire3() => _queue(() async {
-//         print('fire30');
-//         await Future.delayed(const Duration(seconds: 1));
-//         print('fire31');
-//       }, eventId: EventType.event3);
-
-// enum EventType {
-//   event1,
-//   event2,
-//   event3,
-// }
-