+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Add SyncAsyncEventQueue.
 * More flexible queueTransformer.
+* Renamings and restructurings.
+* Improved readme.
 
 ## 0.0.1